+++ packages/storybook/src/nimble/card-button/card-button.stories.ts
@@ -1,13 +1,15 @@
 import { html } from '@microsoft/fast-element';
 import { withActions } from '@storybook/addon-actions/decorator';
 import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
-import { bodyFont } from '@ni/nimble-components/dist/esm/theme-provider/design-tokens';
-import { cardButtonTag } from '@ni/nimble-components/dist/esm/card-button';
-import { createUserSelectedThemeStory } from '../../utilities/storybook';
+import { bodyFont } from '../../../../nimble-components/src/theme-provider/design-tokens';
+import { cardButtonTag } from '../../../../nimble-components/src/card-button';
+import { apiCategory, createUserSelectedThemeStory, disabledDescription } from '../../utilities/storybook';
 
 interface CardButtonArgs {
     disabled: boolean;
     selected: boolean;
+    content: string;
+    click: undefined;
 }
 
 const metadata: Meta<CardButtonArgs> = {
@@ -52,6 +54,28 @@ const metadata: Meta<CardButtonArgs> = {
     args: {
         disabled: false,
         selected: false
+    },
+    argTypes: {
+        disabled: {
+            description: disabledDescription({ componentName: 'card button' }),
+            table: { category: apiCategory.attributes }
+        },
+        selected: {
+            description: 'Styles the card button to indicate it is selected.',
+            table: { category: apiCategory.attributes }
+        },
+        content: {
+            name: 'default',
+            description: 'The card button allows arbitrary HTML child content in its default slot.',
+            table: { category: apiCategory.slots },
+            control: false
+        },
+        click: {
+            description:
+                'Event emitted when the card button is activated by either keyboard or mouse.',
+            table: { category: apiCategory.events },
+            control: false
+        }
     }
 };