+++ packages/storybook/src/nimble/dialog/types.ts
@@ -5,6 +5,13 @@ export const ExampleContentType = {
 export type ExampleContentType =
     (typeof ExampleContentType)[keyof typeof ExampleContentType];
 
+export const ExampleFooterContentType = {
+    text: 'text',
+    buttons: 'buttons'
+} as const;
+export type ExampleFooterContentType =
+    (typeof ExampleFooterContentType)[keyof typeof ExampleFooterContentType];
+
 export const DialogSizeOptions = {
     smallGrowable: 'Small growable',
     largeFixed: 'Large growable'