+++ packages/storybook/src/nimble/menu-button/menu-button.stories.ts
@@ -1,28 +1,30 @@
 import { html, when } from '@microsoft/fast-element';
 import { withActions } from '@storybook/addon-actions/decorator';
 import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
-import { iconArrowExpanderDownTag } from '@ni/nimble-components/dist/esm/icons/arrow-expander-down';
-import { iconKeyTag } from '@ni/nimble-components/dist/esm/icons/key';
-import { menuTag } from '@ni/nimble-components/dist/esm/menu';
-import { anchorMenuItemTag } from '@ni/nimble-components/dist/esm/anchor-menu-item';
-import { menuItemTag } from '@ni/nimble-components/dist/esm/menu-item';
-import { menuButtonTag } from '@ni/nimble-components/dist/esm/menu-button';
+import { iconArrowExpanderDownTag } from '../../../../nimble-components/src/icons/arrow-expander-down';
+import { iconKeyTag } from '../../../../nimble-components/src/icons/key';
+import { menuTag } from '../../../../nimble-components/src/menu';
+import { anchorMenuItemTag } from '../../../../nimble-components/src/anchor-menu-item';
+import { menuItemTag } from '../../../../nimble-components/src/menu-item';
+import { menuButtonTag } from '../../../../nimble-components/src/menu-button';
 import {
     ButtonAppearance,
     ButtonAppearanceVariant,
     MenuButtonPosition
-} from '@ni/nimble-components/dist/esm/menu-button/types';
+} from '../../../../nimble-components/src/menu-button/types';
 import {
     appearanceDescription,
     appearanceVariantDescription,
     contentHiddenDescription,
     endIconDescription,
-    iconDescription
+    iconDescription,
+    textContentDescription
 } from '../patterns/button/doc-strings';
 import {
     apiCategory,
     createUserSelectedThemeStory,
-    disableStorybookZoomTransform
+    disableStorybookZoomTransform,
+    disabledDescription
 } from '../../utilities/storybook';
 
 interface MenuButtonArgs {
@@ -36,8 +38,8 @@ interface MenuButtonArgs {
     disabled: boolean;
     contentHidden: boolean;
     menuPosition: string;
-    toggle: () => void;
-    beforetoggle: () => void;
+    toggle: undefined;
+    beforetoggle: undefined;
 }
 
 const metadata: Meta<MenuButtonArgs> = {
@@ -54,8 +56,7 @@ const metadata: Meta<MenuButtonArgs> = {
     argTypes: {
         label: {
             name: 'default',
-            description:
-                'The text content of the button. This will be hidden when `content-hidden` is set but should always be provided; see the `Accessibility` section for more info.',
+            description: textContentDescription({ componentName: 'menu button' }),
             table: { category: apiCategory.slots }
         },
         icon: {
@@ -94,7 +95,7 @@ const metadata: Meta<MenuButtonArgs> = {
         },
         disabled: {
             control: { type: 'boolean' },
-            description: 'Disables the button.',
+            description: disabledDescription({ componentName: 'menu button' }),
             table: { category: apiCategory.attributes }
         },
         contentHidden: {
@@ -110,13 +111,13 @@ const metadata: Meta<MenuButtonArgs> = {
             table: { category: apiCategory.attributes }
         },
         toggle: {
-            description: 'Fires after the menu button is toggled.',
+            description: 'Event emitted after the menu button is toggled.',
             table: { category: apiCategory.events },
             control: false
         },
         beforetoggle: {
             description:
-                'Fires before the menu button is toggled. This can be used to populate the menu before it is opened.',
+                'Event emitted before the menu button is toggled. This can be used to populate the menu before it is opened.',
             table: { category: apiCategory.events },
             control: false
         }