+++ packages/storybook/src/nimble/table-column/duration-text/table-column-duration-text.stories.ts
@@ -1,16 +1,16 @@
 import { html, ref } from '@microsoft/fast-element';
 import type { HtmlRenderer, Meta, StoryObj } from '@storybook/html';
 import { withActions } from '@storybook/addon-actions/decorator';
-import { tableTag } from '@ni/nimble-components/dist/esm/table';
-import { tableColumnTextTag } from '@ni/nimble-components/dist/esm/table-column/text';
-import { tableColumnDurationTextTag } from '@ni/nimble-components/dist/esm/table-column/duration-text';
+import { tableTag } from '../../../../../nimble-components/src/table';
+import { tableColumnTextTag } from '../../../../../nimble-components/src/table-column/text';
+import { tableColumnDurationTextTag } from '../../../../../nimble-components/src/table-column/duration-text';
 import {
     SharedTableArgs,
     sharedTableActions,
     sharedTableArgTypes,
     sharedTableArgs
 } from '../base/table-column-stories-utils';
-import { createUserSelectedThemeStory } from '../../../utilities/storybook';
+import { apiCategory, createUserSelectedThemeStory } from '../../../utilities/storybook';
 
 const simpleData = [
     {
@@ -95,11 +95,13 @@ export const durationTextColumn: StoryObj<TextColumnTableArgs> = {
             name: 'field-name',
             description:
                 'Set this attribute to identify which field in the data record should be displayed in each column. The field values must be of type `number` and represent a total number of milliseconds.',
-            control: false
+            control: false,
+            table: { category: apiCategory.attributes },
         },
         placeholder: {
             description:
-                'The placeholder text to display when the field value is `undefined` or `null` for a record.'
+                'The placeholder text to display when the field value is `undefined` or `null` for a record.',
+            table: { category: apiCategory.attributes },
         }
     },
     args: {